+++ codec-http/src/main/java/io/netty/handler/codec/http/websocketx/WebSocketProtocolHandler.java
@@ -17,13 +17,20 @@ package io.netty.handler.codec.http.websocketx;
 
 
 import io.netty.channel.ChannelHandlerContext;
+import io.netty.channel.ChannelPromise;
 import io.netty.handler.codec.MessageToMessageDecoder;
+import io.netty.util.ReferenceCountUtil;
+import io.netty.util.concurrent.ScheduledFuture;
 
-import java.util.List;
+import java.nio.channels.ClosedChannelException;
+import java.util.concurrent.TimeUnit;
 
 abstract class WebSocketProtocolHandler extends MessageToMessageDecoder<WebSocketFrame> {
 
     private final boolean dropPongFrames;
+    private final WebSocketCloseStatus closeStatus;
+    private final long forceCloseTimeoutMillis;
+    private ChannelPromise closeSent;
 
     /**
      * Creates a new {@link WebSocketProtocolHandler} that will <i>drop</i> {@link PongWebSocketFrame}s.
@@ -40,7 +47,15 @@ abstract class WebSocketProtocolHandler extends MessageToMessageDecoder<WebSocke
      *            {@code true} if {@link PongWebSocketFrame}s should be dropped
      */
     WebSocketProtocolHandler(boolean dropPongFrames) {
+        this(dropPongFrames, null, 0L);
+    }
+
+    WebSocketProtocolHandler(boolean dropPongFrames,
+                             WebSocketCloseStatus closeStatus,
+                             long forceCloseTimeoutMillis) {
         this.dropPongFrames = dropPongFrames;
+        this.closeStatus = closeStatus;
+        this.forceCloseTimeoutMillis = forceCloseTimeoutMillis;
     }
 
     @Override
@@ -63,6 +78,48 @@ abstract class WebSocketProtocolHandler extends MessageToMessageDecoder<WebSocke
         if (!ctx.channel().config().isAutoRead()) {
             ctx.read();
         }
+    }
+
+    @Override
+    public void close(final ChannelHandlerContext ctx, final ChannelPromise promise) throws Exception {
+        if (closeStatus == null || !ctx.channel().isActive()) {
+            ctx.close(promise);
+        } else {
+            if (closeSent == null) {
+                write(ctx, new CloseWebSocketFrame(closeStatus), ctx.newPromise());
+            }
+            flush(ctx);
+            applyCloseSentTimeout(ctx);
+            closeSent.addListener(future -> ctx.close(promise));
+        }
+    }
+
+    @Override
+    public void write(final ChannelHandlerContext ctx, Object msg, ChannelPromise promise) throws Exception {
+        if (closeSent != null) {
+            ReferenceCountUtil.release(msg);
+            promise.setFailure(new ClosedChannelException());
+            return;
+        }
+        if (msg instanceof CloseWebSocketFrame) {
+            promise = promise.unvoid();
+            closeSent = promise;
+        }
+        ctx.write(msg, promise);
+    }
+
+    private void applyCloseSentTimeout(ChannelHandlerContext ctx) {
+        if (closeSent.isDone() || forceCloseTimeoutMillis < 0) {
+            return;
+        }
+
+        final ScheduledFuture<?> timeoutTask = ctx.executor().schedule(() -> {
+            if (!closeSent.isDone()) {
+                closeSent.tryFailure(new WebSocketHandshakeException("send close frame timed out"));
+            }
+        }, forceCloseTimeoutMillis, TimeUnit.MILLISECONDS);
+
+        closeSent.addListener(future -> timeoutTask.cancel(false));
     }
 
     @Override
